addressed fixes requested by Remi July 3rd
