+++ pdns/dnsdist-lua.cc
@@ -1586,54 +1586,47 @@ vector<std::function<void(void)>> setupLua(bool client, const std::string& confi
       }
     });
 
-    g_lua.registerFunction<void(DNSQuestion::*)(std::string, std::string)>("setTag", [](DNSQuestion& dq, const std::string& strLabel, const std::string& strValue) {
+  g_lua.registerFunction<void(DNSQuestion::*)(std::string, std::string)>("setTag", [](DNSQuestion& dq, const std::string& strLabel, const std::string& strValue) {
 
-      if(dq.qTag == NULL)
-        {
-         dq.qTag = std::shared_ptr<QTag>(new QTag);
-        }
+      if(dq.qTag == nullptr) {
+        dq.qTag = std::make_shared<QTag>();
+      }
       dq.qTag->add(strLabel, strValue);
 
     });
 
-     g_lua.registerFunction<void(DNSQuestion::*)(vector<pair<string, string>>)>("setTagArray", [](DNSQuestion& dq, const vector<pair<string, string>>&tags) {
+  g_lua.registerFunction<void(DNSQuestion::*)(vector<pair<string, string>>)>("setTagArray", [](DNSQuestion& dq, const vector<pair<string, string>>&tags) {
 
-      if(dq.qTag == NULL)
-        {
-         dq.qTag = std::shared_ptr<QTag>(new QTag);
-        }
+      if(dq.qTag == nullptr) {
+        dq.qTag = std::make_shared<QTag>();
+      }
 
-      for (const auto& tag : tags)
-        {
-           dq.qTag->add(tag.first, tag.second);
-        }
+      for (const auto& tag : tags) {
+        dq.qTag->add(tag.first, tag.second);
+      }
 
      });
 
 
-     g_lua.registerFunction<string(DNSQuestion::*)(std::string)>("getTagMatch", [](const DNSQuestion& dq, const std::string& strLabel) {
+  g_lua.registerFunction<string(DNSQuestion::*)(std::string)>("getTag", [](const DNSQuestion& dq, const std::string& strLabel) {
 
     std::string strValue;
-    if(dq.qTag != NULL)
-      {
-        strValue = dq.qTag->getMatch(strLabel);
-      }
+    if(dq.qTag != nullptr) {
+      strValue = dq.qTag->getMatch(strLabel);
+    }
     return strValue;
 
      });
 
 
-     g_lua.registerFunction<std::unordered_map<string, string>(DNSQuestion::*)(void)>("getTagArray", [](const DNSQuestion& dq) {
+  g_lua.registerFunction<std::unordered_map<string, string>(DNSQuestion::*)(void)>("getTagArray", [](const DNSQuestion& dq) {
 
-    if(dq.qTag != NULL)
-      {
-        return dq.qTag->tagData;
-      }
-    else
-      {
-       std::unordered_map<string, string> XX;
-       return(XX);
-      }
+    if(dq.qTag != nullptr) {
+      return dq.qTag->tagData;
+    } else {
+      std::unordered_map<string, string> XX;
+      return XX;
+    }
       });
