+++ pdns/protobuf.cc
@@ -110,7 +110,7 @@ void DNSProtoBufMessage::addTag(const std::string& strValue)
 #endif /* HAVE_PROTOBUF */
 }
 
-void DNSProtoBufMessage::addRR(const std::string& strName, uint32_t uType, uint32_t uClass, uint32_t uTTL, const uint8_t *ptrBlob, size_t uBlobLen)
+void DNSProtoBufMessage::addRR(const std::string& strName, uint32_t uType, uint32_t uClass, uint32_t uTTL, const std::string& strBlob)
 {
 #ifdef HAVE_PROTOBUF
 
@@ -120,12 +120,11 @@ void DNSProtoBufMessage::addRR(const std::string& strName, uint32_t uType, uint3
 
   PBDNSMessage_DNSResponse_DNSRR* rr = response->add_rrs();
       if (rr) {
-        string blob;
         rr->set_name(strName.c_str());
         rr->set_type(uType);
         rr->set_class_(uClass);
         rr->set_ttl(uTTL);
-        rr->set_rdata(ptrBlob, uBlobLen);
+        rr->set_rdata((const uint8_t *) strBlob.c_str(), strBlob.size());
       }
 
 #endif /* HAVE_PROTOBUF */
